pyproject.toml
[project]
name = "fastapi_backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.1",
    "fastapi-users[sqlalchemy]>=13.0.0",
    "httpx>=0.28.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.0.0",
    "pwdlib[argon2]>=0.3.0",
    "python-jose>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.45",
]

# 可选依赖分组 - 专门存放测试/开发依赖（生产环境不会主动安装）
[project.optional-dependencies]

# dev 额外包（用于容器/开发环境），复制自 test，以便 `uv sync --extra dev` 生效
dev = [
    "python-dotenv>=1.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
